Remove gateway properties in SystemConfiguration
Now default gateway are per-interface properties, the system-wide one
in the SystemConfiguration interface is deprecated. Updating the value
also makes no difference, it will never be written to configuration.
This patch removes the code of initializing and updating them.

Change #52785 removes the properties in phosphor-dbus-interfaces yaml,
it is recommended to be merged after this change.

Tested:
In SystemConfiguration interface, DefaultGateway and DefaultGateway6 are
still there, but writing them only updates the displayed value on DBus.
No code in phosphor-networkd is called.

Change-Id: I58ac52c5318943941dcc92abede7763bddbca596
Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>