ethernet_interface: Support DHCP4/6 Properties
We are splitting the DHCPEnabled option into 3 new options to reflect
all of the desired states (DHCP4, DHCP6, IPv6AcceptRA). We will
still support DHCPEnabled until all of the old users have transitioned
away. This commit makes DHCPEnabled a view / modifier of the other
options.

Tested: Toggled the DHCPEnabled settings on a BMC and verified that the
state updated correctly along with the other properties to affects.
Verified that changing the individual properties updates the DHCPEnabled
property to the correct enum state.

Change-Id: I2cef15d7eaf80c8d717a11d090f61ca5d275821a
Signed-off-by: William A. Kennington III <wak@google.com>