configuration - add DHCP vendor classification identification
Vendor class identification in DHCP is a mechanism used to identify
the type or class of DHCP client making a request. This allows DHCP
servers to tailor their responses based on the specific requirements
or characteristics of different types of clients.

The DHCP protocol allows clients to include a "vendor class identifier"
(option 60) in their DHCP discover or request packets. This identifier
typically contains a string that uniquely identifies the type or
class of the client

This commit adds a configuration option for the network interfaces to
specify the DHCP vendor class identification.

fixes nvbug https://redmine.mellanox.com/issues/3898337