Handle IPv4 address for IPv6 Static Default Gateway
Error is not thrown when an IPV4 address is provided.
This commit addresses the issue by explicitly handling such cases
and returning an InvalidArgument error when a IPV4address
is configured for IPV6 Static Default Gateway field.

Tested By:
Verified the test case and ensured proper
invalid argument error is thrown.

Change-Id: Ibc3d2a77ae56f5b23c3f8bca12409693aa2f3ad9
Signed-off-by: kokilav <kokilavaradhan@gmail.com>