OWNERS: add Raviteja as an owner
I have worked on various features and fixed several issues in this
Networkd repository and redfish EthernetInterface service

There are still some patches that need to be reviewed by the
community, and some are ready to be merged

I have thorough understanding of Networkd and systemd-networkd
I am also active reviewer of this repo from last two years
Thus, I apply to be a maintainer of this repo. Hope I can contribute
more and help the OpenBMC community.

Here are my contributions to phosphor-networkd repository [1]:
[1]: https://github.com/openbmc/phosphor-networkd/commits?author=raviteja-b

Change-Id: I7fe5364a45140356617e1b72cd949a2e28c46a42
Signed-off-by: Ravi Teja <raviteja28031990@gmail.com>